[ci skip] Update TaskSerializer - Move code main to cpp file
